refactor #51: utils 경로 설정 및 파일 추가
- getTimeElapsed 추가
