+++ fe/tsconfig.json
@@ -20,7 +20,8 @@
       "@common/*": ["src/common/*"],
       "@components/*": ["src/components/*"],
       "@assets/*": ["src/assets/*"],
-      "@constants/*": ["src/constants/*"]
+      "@constants/*": ["src/constants/*"],
+      "@utils/*": ["src/utils/*"],
     }
   },
   "include": ["src", "craco.config.js"],
