+++ fe/tsconfig.paths.json
@@ -5,7 +5,8 @@
       "@common/*": ["src/common/*"],
       "@components/*": ["src/components/*"],
       "@assets/*": ["src/assets/*"],
-      "@constants/*": ["src/constants/*"]
+      "@constants/*": ["src/constants/*"],
+      "@utils/*": ["src/utils/*"]
     }
   }
 }
